+++ craft_parts/executor/executor.py
@@ -16,8 +16,10 @@
 
 """Definitions and helpers for the action executor."""
 
+import itertools
 import logging
 import shutil
+from collections.abc import Iterable, Iterator
 from pathlib import Path
 
 from typing_extensions import Self
@@ -55,6 +57,7 @@ class Executor:
     :param ignore_patterns: File patterns to ignore when pulling local sources.
     :param use_host_sources: Whether overlay steps should also include the repository
       sources defined on the host.
+    :param build_environment: The environment variables to be set during build.
     """
 
     def __init__(  # noqa: PLR0913
@@ -69,6 +72,7 @@ class Executor:
         base_layer_dir: Path | None = None,
         base_layer_hash: LayerHash | None = None,
         use_host_sources: bool = False,
+        build_environment: Iterable[str] | None = None,
     ) -> None:
         self._part_list = sort_parts(part_list)
         self._project_info = project_info
@@ -79,6 +83,7 @@ class Executor:
         self._handler: dict[str, PartHandler] = {}
         self._ignore_patterns = ignore_patterns
         self._use_host_sources = use_host_sources
+        self._build_environment = build_environment
 
         # The cache layer level is set to the first part that doesn't organize
         # to the overlay coming after a part that organizes to the overlay.
@@ -260,6 +265,12 @@ class Executor:
         if part.name in self._handler:
             return self._handler[part.name]
 
+        build_environment = self._build_environment
+        if isinstance(self._build_environment, Iterator):
+            # Give a new generator instance to each part
+            build_environment, next_gen = itertools.tee(self._build_environment, 2)
+            self._build_environment = next_gen
+
         handler = PartHandler(
             part,
             part_info=PartInfo(self._project_info, part),
@@ -268,6 +279,7 @@ class Executor:
             overlay_manager=self._overlay_manager,
             ignore_patterns=self._ignore_patterns,
             base_layer_hash=self._base_layer_hash,
+            build_environment=build_environment,
         )
         self._handler[part.name] = handler